requirements.txt
Django==1.11rc1
gunicorn==19.7.1
django-multi-email-field==0.4
